Handle MMS without a MessageType.
Added handling of MMS messages without a MessageType specified.  In
order to handle this, we just assume that the lack of a MessageType
means that the message was sent to yourself.  This assumption allows us
to handle it by simply storing a sent and a received version of the
message.